Refactored Statements and Expressions by adding Generate and GenerateExpr traits
