+++ src/codegen/expressions/expr_list.rs
@@ -1,0 +1,7 @@
+use crate::parser::expressions::expr_list::ExpressionList;
+
+use super::GenerateExpr;
+
+impl GenerateExpr for ExpressionList {
+    // Not needed
+}
