+++ src/codegen/statements/fors.rs
@@ -1,0 +1,7 @@
+use crate::parser::statements::fors::ForStatement;
+
+use super::Generate;
+
+impl Generate for ForStatement { 
+    // Not needed
+}
