+++ src/parser/expressions/literals.rs
@@ -1,5 +1,3 @@
-use inkwell::values::AnyValue;
-
 use super::{AnyExpressionEnum, Expression};
 use super::expr_list::ExpressionList;
 #[derive(Debug, Clone)]
@@ -25,12 +23,6 @@ pub struct NumberLiteral {
     pub val: String,
 }
 impl Expression for NumberLiteral {
-    fn codegen_expression<'ctx>(&self, codegen: &'ctx crate::codegen::CodeGen) -> inkwell::values::AnyValueEnum<'ctx>{
-        let f = self.val.split(' ').filter_map(|s| s.parse::<f32>().ok()).collect::<Vec<_>>();
-        let f = *f.first().unwrap() as f64;
-
-        codegen.context.f64_type().const_float(f).as_any_value_enum()
-    }
     fn into_any_expression_enum(self) -> super::AnyExpressionEnum {
         super::AnyExpressionEnum::NumberLiteral(self)
     }
@@ -70,9 +62,6 @@ impl Expression for False {
     fn into_any_expression_enum(self) -> super::AnyExpressionEnum {
         super::AnyExpressionEnum::False(self)
     }
-    fn codegen_expression<'ctx>(&self, codegen: &'ctx crate::codegen::CodeGen) -> inkwell::values::AnyValueEnum<'ctx>{
-        codegen.context.bool_type().const_zero().as_any_value_enum()
-    }
     fn desugar(self) -> AnyExpressionEnum {
         self.clone().into_any_expression_enum()
     }
@@ -86,9 +75,6 @@ pub struct True;
 impl Expression for True {
     fn into_any_expression_enum(self) -> super::AnyExpressionEnum {
         super::AnyExpressionEnum::True(self)
-    }
-    fn codegen_expression<'ctx>(&self, codegen: &'ctx crate::codegen::CodeGen) -> inkwell::values::AnyValueEnum<'ctx>{
-        codegen.context.bool_type().const_all_ones().as_any_value_enum()
     }
     fn desugar(self) -> AnyExpressionEnum {
         self.clone().into_any_expression_enum()
