+++ src/parser/expressions/unary.rs
@@ -1,6 +1,3 @@
-
-use inkwell::values::AnyValue;
-
 use super::{AnyExpressionEnum, Expression};
 #[derive(Debug)]
 pub struct UnaryOp {
@@ -10,25 +7,6 @@ pub struct UnaryOp {
 
 
 impl Expression for UnaryOp {
-    fn codegen_expression<'ctx>(&self, codegen: &'ctx crate::codegen::CodeGen) -> inkwell::values::AnyValueEnum<'ctx> {
-        let v = self.operand.codegen_expression(codegen);
-        match self.op_type {
-            UnaryOpType::Not => {
-                let v = v.into_int_value();
-                codegen.builder
-                    .build_not(v, "")
-                    .unwrap()
-                    .as_any_value_enum()
-            }
-            UnaryOpType::Minus => {
-                let v = v.into_float_value();
-                codegen.builder
-                    .build_float_neg(v, "")
-                    .unwrap()
-                    .as_any_value_enum()
-            }
-        }
-    }
     fn into_any_expression_enum(self) -> AnyExpressionEnum {
         AnyExpressionEnum::Unary(self)
     }
